// atlassian-frontend/packages/dragonfruit/search-dialog/src/confluence/clients/response-types.tsx
export const GenericContainerResult: 'generic-container-result' = 'generic-container-result' as const;
export const PersonResult: 'person-result' = 'person-result' as const;
export const ConfluenceObjectResult: 'confluence-object-result' = 'confluence-object-result' as const;

export type ResultType =
  | typeof GenericContainerResult
  | typeof PersonResult
  | typeof ConfluenceObjectResult;

/********************/
/** Analytics Type **/
/********************/
export const RecentConfluence: 'recent-confluence' = 'recent-confluence' as const;
export const ResultConfluence: 'result-confluence' = 'result-confluence' as const;
export const RecentPerson: 'recent-person' = 'recent-person' as const;
export const ResultPerson: 'result-person' = 'result-person' as const;
// ...
export const ConfluencePage: 'confluence-page' = 'confluence-page' as const;
export const ConfluenceBlogpost: 'confluence-blogpost' = 'confluence-blogpost' as const;
export const ConfluenceAttachment: 'confluence-attachment' = 'confluence-attachment' as const;
export const ConfluenceSpace: 'confluence-space' = 'confluence-space' as const;
export const Person: 'person' = 'person' as const;

export type ContentType =
  | typeof ConfluencePage
  | typeof ConfluenceBlogpost
  | typeof ConfluenceAttachment
  | typeof ConfluenceSpace
  | typeof Person;

export interface Result {
  resultId: string;
  // main text to show
  name: string;
  // url to link the result to
  href: string;
  // the analytics type to send in the analytics attributes
  analyticsType: AnalyticsType;
  // field to disambiguate between result types
  resultType: ResultType;
  contentType: ContentType;
}

export interface Results<T extends Result> {
  items: T[];
  timings: number;
}
// ...
export interface ConfPeopleResult extends Result {
  mentionName: string;
  // the message to display underneath the name, unfortuntately named this way ATM.
  presenceMessage: string;
  resultType: typeof PersonResult;
  avatarUrl: string;
  contentType: typeof Person;
  userId: string;
}

export interface ConfPeopleResults extends Results<ConfPeopleResult> {}
// ...
export interface ConfItemResult extends Result {
  containerName: string;
  containerId: string;
  resultType: typeof ConfluenceObjectResult;
  iconClass?: string;
  lastModified: string | undefined;
  // used to indicate the result came from the recently viewed FE cache
  isRecentResult: boolean;
}

export interface ConfItemResults extends Results<ConfItemResult> {
  totalSize: number;
}

export interface ConfSpaceResponse
  extends ScopedAggregatorResponse<Scope.ConfluenceSpace> {
  results: ConfItemResponseItem[];
}

export interface ConfSpaceResult extends Result {
  resultType: typeof GenericContainerResult;
  contentType: typeof ConfluenceSpace;
  avatarUrl: string;
  key: string;
}

export interface ConfSpaceResults extends Results<ConfSpaceResult> {}